last comittment before big correction
process函数之后的so3数据是对的；但是返回的四元数是有问题的。两种方案：直接用process函数中的结果；找出返回阶段的错误。

+++ src/utils/HeadTransform.java
@@ -1,36 +1,38 @@
 package utils;
 
+import java.math.BigDecimal;
+
 public class HeadTransform {
     private static final double GIMBAL_LOCK_EPSILON = 0.01f;
-    private final double[] mHeadView;
+    private final float[] mHeadView;
 
     public HeadTransform() {
         super();
-        Matrix.setIdentityM(this.mHeadView = new double[16], 0);
+        Matrix.setIdentityM(this.mHeadView = new float[16], 0);
     }
 
-    public double[] getHeadView() {
+    public float[] getHeadView() {
         return this.mHeadView;
     }
 
-    public void setmHeadView(double[] mHeadView){
+    public void setmHeadView(float[] mHeadView){
         for(int i=0;i<mHeadView.length;i++){
             this.mHeadView[i] = mHeadView[i];
         }
     }
 
-    public void getQuaternion(final double[] quaternion, final int offset) {
+    public void getQuaternion(final float[] quaternion, final int offset) {
         if (offset + 4 > quaternion.length) {
             throw new IllegalArgumentException("Not enough space to write the result");
         }
-        final double[] m = this.mHeadView;
-        final double t = m[0] + m[5] + m[10];
-        double w;
-        double x;
-        double y;
-        double z;
+        final float[] m = this.mHeadView;
+        final float t = m[0] + m[5] + m[10];
+        float w;
+        float x;
+        float y;
+        float z;
         if (t >= 0.0f) {
-            double s = Math.sqrt(t + 1.0f);
+            float s = (float)Math.sqrt(t+1.0f);
             w = 0.5f * s;
             s = 0.5f / s;
             x = (m[9] - m[6]) * s;
@@ -38,7 +40,7 @@ public class HeadTransform {
             z = (m[4] - m[1]) * s;
         }
         else if (m[0] > m[5] && m[0] > m[10]) {
-            double s = Math.sqrt(1.0f + m[0] - m[5] - m[10]);
+            float s = (float)Math.sqrt(1.0f + m[0] - m[5] - m[10]);
             x = s * 0.5f;
             s = 0.5f / s;
             y = (m[4] + m[1]) * s;
@@ -46,7 +48,7 @@ public class HeadTransform {
             w = (m[9] - m[6]) * s;
         }
         else if (m[5] > m[10]) {
-            double s = Math.sqrt(1.0f + m[5] - m[0] - m[10]);
+            float s = (float) Math.sqrt(1.0f + m[5] - m[0] - m[10]);
             y = s * 0.5f;
             s = 0.5f / s;
             x = (m[4] + m[1]) * s;
@@ -54,7 +56,7 @@ public class HeadTransform {
             w = (m[2] - m[8]) * s;
         }
         else {
-            double s = Math.sqrt(1.0f + m[10] - m[0] - m[5]);
+            float s = (float) Math.sqrt(1.0f + m[10] - m[0] - m[5]);
             z = s * 0.5f;
             s = 0.5f / s;
             x = (m[2] + m[8]) * s;
